+++ browser/android/src/main/java/com/capacitorjs/plugins/browser/BrowserPlugin.java
@@ -60,25 +60,23 @@ public class BrowserPlugin extends Plugin {
         }
 
         // open the browser and finish
-        try {
-            Intent intent = new Intent(getContext(), BrowserControllerActivity.class);
-            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            getContext().startActivity(intent);
 
-            Integer finalToolbarColor = toolbarColor;
-            setBrowserControllerListener(
-                activity -> {
-                    activity.open(implementation, url, finalToolbarColor);
-                    browserControllerActivityInstance = activity;
-                    call.resolve();
-                }
-            );
-        } catch (ActivityNotFoundException ex) {
-            Logger.error(getLogTag(), ex.getLocalizedMessage(), null);
-            call.reject("Unable to display URL");
-            return;
-        }
+        Intent intent = new Intent(getContext(), BrowserControllerActivity.class);
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        getContext().startActivity(intent);
+
+        Integer finalToolbarColor = toolbarColor;
+        setBrowserControllerListener((activity) -> {
+            try {
+                activity.open(implementation, url, finalToolbarColor);
+                browserControllerActivityInstance = activity;
+                call.resolve();
+            } catch (ActivityNotFoundException ex) {
+                Logger.error(getLogTag(), ex.getLocalizedMessage(), null);
+                call.reject("Unable to display URL");
+            }
+        });
     }
 
     @PluginMethod