+++ camera/src/web.ts
@@ -52,15 +52,10 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
     });
   }
 
-  private async cameraExperience(
-    options: ImageOptions,
-    resolve: any,
-    reject: any,
-  ) {
+  private async cameraExperience(options: ImageOptions, resolve: any, reject: any) {
     if (customElements.get('pwa-camera-modal')) {
       const cameraModal: any = document.createElement('pwa-camera-modal');
-      cameraModal.facingMode =
-        options.direction === CameraDirection.Front ? 'user' : 'environment';
+      cameraModal.facingMode = options.direction === CameraDirection.Front ? 'user' : 'environment';
       document.body.appendChild(cameraModal);
       try {
         await cameraModal.componentOnReady();
@@ -91,14 +86,8 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
     }
   }
 
-  private fileInputExperience(
-    options: ImageOptions,
-    resolve: any,
-    reject: any,
-  ) {
-    let input = document.querySelector(
-      '#_capacitor-camera-input',
-    ) as HTMLInputElement;
+  private fileInputExperience(options: ImageOptions, resolve: any, reject: any) {
+    let input = document.querySelector('#_capacitor-camera-input') as HTMLInputElement;
 
     const cleanup = () => {
       input.parentNode?.removeChild(input);
@@ -120,10 +109,7 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
           format = 'gif';
         }
 
-        if (
-          options.resultType === 'dataUrl' ||
-          options.resultType === 'base64'
-        ) {
+        if (options.resultType === 'dataUrl' || options.resultType === 'base64') {
           const reader = new FileReader();
 
           reader.addEventListener('load', () => {
@@ -161,10 +147,7 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
     input.accept = 'image/*';
     (input as any).capture = true;
 
-    if (
-      options.source === CameraSource.Photos ||
-      options.source === CameraSource.Prompt
-    ) {
+    if (options.source === CameraSource.Photos || options.source === CameraSource.Prompt) {
       input.removeAttribute('capture');
     } else if (options.direction === CameraDirection.Front) {
       (input as any).capture = 'user';
@@ -176,9 +159,7 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
   }
 
   private multipleFileInputExperience(resolve: any, reject: any) {
-    let input = document.querySelector(
-      '#_capacitor-camera-input-multiple',
-    ) as HTMLInputElement;
+    let input = document.querySelector('#_capacitor-camera-input-multiple') as HTMLInputElement;
 
     const cleanup = () => {
       input.parentNode?.removeChild(input);
@@ -250,7 +231,7 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
             });
           }
         };
-        reader.onerror = e => {
+        reader.onerror = (e) => {
           reject(e);
         };
       }
@@ -274,9 +255,7 @@ export class CameraWeb extends WebPlugin implements CameraPlugin {
         photos: 'granted',
       };
     } catch {
-      throw this.unavailable(
-        'Camera permissions are not available in this browser',
-      );
+      throw this.unavailable('Camera permissions are not available in this browser');
     }
   }