+++ preferences/src/index.ts
@@ -3,7 +3,7 @@ import { registerPlugin } from '@capacitor/core';
 import type { PreferencesPlugin } from './definitions';
 
 const Preferences = registerPlugin<PreferencesPlugin>('Preferences', {
-  web: () => import('./web').then(m => new m.PreferencesWeb()),
+  web: () => import('./web').then((m) => new m.PreferencesWeb()),
 });
 
 export * from './definitions';