+++ push-notifications/Package.swift
@@ -3,14 +3,14 @@ import PackageDescription
 
 let package = Package(
     name: "CapacitorPushNotifications",
-    platforms: [.iOS(.v14)],
+    platforms: [.iOS(.v15)],
     products: [
         .library(
             name: "CapacitorPushNotifications",
             targets: ["PushNotificationsPlugin"])
     ],
     dependencies: [
-        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", branch: "main")
+        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", from: "8.0.0")
     ],
     targets: [
         .target(