+++ status-bar/android/src/main/java/com/capacitorjs/plugins/statusbar/StatusBarPlugin.java
@@ -1,5 +1,6 @@
 package com.capacitorjs.plugins.statusbar;
 
+import android.content.res.Configuration;
 import com.getcapacitor.JSObject;
 import com.getcapacitor.Logger;
 import com.getcapacitor.Plugin;
@@ -49,6 +50,12 @@ public class StatusBarPlugin extends Plugin {
         }
     }
 
+    @Override
+    protected void handleOnConfigurationChanged(Configuration newConfig) {
+        super.handleOnConfigurationChanged(newConfig);
+        implementation.updateStyle();
+    }
+
     @PluginMethod
     public void setStyle(final PluginCall call) {
         final String style = call.getString("style");
@@ -57,13 +64,10 @@ public class StatusBarPlugin extends Plugin {
             return;
         }
 
-        getBridge()
-            .executeOnMainThread(
-                () -> {
-                    implementation.setStyle(style);
-                    call.resolve();
-                }
-            );
+        getBridge().executeOnMainThread(() -> {
+            implementation.setStyle(style);
+            call.resolve();
+        });
     }
 
     @PluginMethod
@@ -74,42 +78,33 @@ public class StatusBarPlugin extends Plugin {
             return;
         }
 
-        getBridge()
-            .executeOnMainThread(
-                () -> {
-                    try {
-                        final int parsedColor = WebColor.parseColor(color.toUpperCase(Locale.ROOT));
-                        implementation.setBackgroundColor(parsedColor);
-                        call.resolve();
-                    } catch (IllegalArgumentException ex) {
-                        call.reject("Invalid color provided. Must be a hex string (ex: #ff0000");
-                    }
-                }
-            );
+        getBridge().executeOnMainThread(() -> {
+            try {
+                final int parsedColor = WebColor.parseColor(color.toUpperCase(Locale.ROOT));
+                implementation.setBackgroundColor(parsedColor);
+                call.resolve();
+            } catch (IllegalArgumentException ex) {
+                call.reject("Invalid color provided. Must be a hex string (ex: #ff0000");
+            }
+        });
     }
 
     @PluginMethod
     public void hide(final PluginCall call) {
         // Hide the status bar.
-        getBridge()
-            .executeOnMainThread(
-                () -> {
-                    implementation.hide();
-                    call.resolve();
-                }
-            );
+        getBridge().executeOnMainThread(() -> {
+            implementation.hide();
+            call.resolve();
+        });
     }
 
     @PluginMethod
     public void show(final PluginCall call) {
         // Show the status bar.
-        getBridge()
-            .executeOnMainThread(
-                () -> {
-                    implementation.show();
-                    call.resolve();
-                }
-            );
+        getBridge().executeOnMainThread(() -> {
+            implementation.show();
+            call.resolve();
+        });
     }
 
     @PluginMethod
@@ -121,13 +116,10 @@ public class StatusBarPlugin extends Plugin {
     @PluginMethod
     public void setOverlaysWebView(final PluginCall call) {
         final Boolean overlay = call.getBoolean("overlay", true);
-        getBridge()
-            .executeOnMainThread(
-                () -> {
-                    implementation.setOverlaysWebView(overlay);
-                    call.resolve();
-                }
-            );
+        getBridge().executeOnMainThread(() -> {
+            implementation.setOverlaysWebView(overlay);
+            call.resolve();
+        });
     }
 
     private JSObject toJSObject(StatusBarInfo info) {